+++ project/packages/storybook/jest-puppeteer.config.js
@@ -1,0 +1,6 @@
+module.exports = {
+  server: {
+    command: `NODE_ENV="development" && npm run start`,
+    port: 6006,
+  },
+};